refactored get trailer registration to trailer repository
